Solution to Simple Calculator

+++ Simple Calculator@@6788dd63490637cfc44e1ba8/code.c
@@ -1,15 +1,15 @@
 #include<stdio.h>
 int main(){
-    double num1 ,num2, result;
+    int num1 ,num2, result;
     char op;
-    scanf("%lf %lf %c",&num1,&num2,&op);
+    scanf("%d %d %c",&num1,&num2,&op);
     if(op == '+' ) result =num1 + num2;
     else if(op == '*')result = num1 * num2;
     else if(op == '-')result = num1 - num2;
     else if(op == '/')result = num1 / num2;
     else return printf("error");
 
-    printf("%.2lf\n",result);
+    printf("%d\n",result);
     return 0;
